fix(hunyuan-video10): scale sigma in fp32 before the transformer cast
HV1.0's predict_noise cast sigma to the transformer dtype and only then
multiplied by 1000, so the bf16 rounding happened on the [0,1] sigma
rather than on the [0,1000] timestep. Diffusers and the SGLang scheduler
both scale in fp32 and cast afterwards, so the trainside replay fed the
DiT different timesteps than the rollout engine on 2 of the 10 recipe
steps. PR #390 made exactly this change for HV1.5 and left HV1.0 alone.

+++ unirl/models/hunyuan_video10/diffusion.py
@@ -60,7 +60,8 @@ class HunyuanVideo10DiffusionStep(DiffusionStep[HunyuanVideo10Bundle, HunyuanVid
             timestep = sigma.expand(batch_size)
         else:
             timestep = sigma
-        timestep = timestep.to(device=device, dtype=dtype) * self.TIMESTEP_SCALE
+        # Match Diffusers/SGLang: scale fp32 sigma before casting to the transformer dtype.
+        timestep = (timestep.to(device=device, dtype=torch.float32) * self.TIMESTEP_SCALE).to(dtype=dtype)
 
         guidance = torch.full((batch_size,), guidance_scale, device=device, dtype=dtype)
